fix(RateLimiter): fail-closed for unconfigured issuers (#51)
* fix(RateLimiter): fail-closed for unconfigured issuers

Unconfigured issuers previously had unlimited mint capacity (silent
pass-through). Now _enforceRateLimit reverts with RateLimitNotConfigured,
view functions return zero, and the deploy script bootstraps the
IssuerGateway rate limit atomically with transient role management.

Sherlock L-04

* fix(RateLimiter): guard zero-capacity config and bootstrap earnerManager rate limit

- Add InvalidRateLimitConfig error and guard against enabled=true with capacity=0,
  preventing a "configured but always reverts" bucket state
- Bootstrap earnerManager with unlimited capacity so distributeReward() flows
  through _mint without hitting RateLimitNotConfigured

+++ test/unit/PYUSDXUnit.t.sol
@@ -242,8 +242,8 @@ contract PYUSDXUnitTests is PYUSDXBaseUnitTest {
     }
 
     function test_mint_maxSafeAmount() public {
-        // The maximum safe amount is limited by uint240 totalSupply
-        uint240 maxSafeAmount = type(uint240).max;
+        // The maximum safe amount is limited by uint128 rate-limit capacity
+        uint128 maxSafeAmount = type(uint128).max;
 
         issuerGateway.mint(alice, uint256(maxSafeAmount));
